requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Authentication
Authlib==1.3.0
cryptography==41.0.7

# Database
SQLAlchemy==2.0.23

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Machine Learning (Intent Recognition)
torch==2.1.2
numpy==1.26.2
scikit-learn==1.3.2

# WSGI Server (Production)
gunicorn==21.2.0

# Browser Automation & Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# OpenAI (for GPT-5-mini vision)
openai==1.10.0

# Database Drivers
psycopg2-binary==2.9.9  # PostgreSQL
PyMuPDF
